fix: cast checkpoint reward to native float for JSON serialization
_policy_mean_reward returned numpy.float32 (from summing episode
rewards), which crashed training_status.json's save_state with
"TypeError: Object of type float32 is not JSON serializable" whenever
a checkpoint was written.

+++ ml-agents/mlagents/trainers/trainer/rl_trainer.py
@@ -128,7 +128,8 @@ class RLTrainer(Trainer):
         if len(rewards) == 0:
             return None
         else:
-            return sum(rewards) / len(rewards)
+            # Cast from numpy.float32 (non-JSON-serializable) to a native float
+            return float(sum(rewards) / len(rewards))
 
     @timed
     def _checkpoint(self) -> ModelCheckpoint:
